Moving files to better organize project
