+++ src/controllers/test-setup.js
@@ -41,51 +41,3 @@ module.exports = async function(req, res) {
     }
   })
 }
-
-/**
- * "data": {
-    "samples": {
-      "triggers": {
-        "geo_location": {
-          "location": {
-            "lat": "",
-            "lon": "",
-            "address": "",
-            "description": "",
-            "radius": ""
-          }
-        },
-        "open_data_cd01961": {
-          "dataset": ""
-        },
-        "open_data": {
-          "dataset": ""
-        },
-        "alberta_air_health_risk": {
-          "city": ""
-        },
-        "alberta_air_health_index": {
-          "city": ""
-        }
-      },
-      "actions": {
-        "notification": {
-          "notification": ""
-        },
-        "test": {
-          "test": "",
-          "test2": ""
-        }
-      },
-      "actionRecordSkipping": {
-        "notification": {
-          "notification": ""
-        },
-        "test": {
-          "test": "",
-          "test2": ""
-        }
-      }
-    }
-  }
- */
